+++ sdk/src/main/java/com/mapbox/maps/renderer/MapboxSurfaceRenderer.kt
@@ -10,23 +10,30 @@ internal open class MapboxSurfaceRenderer : MapboxRenderer {
 
   override val widgetRenderer: MapboxWidgetRenderer
 
-  constructor(antialiasingSampleCount: Int, contextMode: ContextMode) {
+  constructor(
+    antialiasingSampleCount: Int,
+    contextMode: ContextMode,
+    mapName: String,
+  ) : super(mapName) {
     widgetRenderer = MapboxWidgetRenderer(
       antialiasingSampleCount = antialiasingSampleCount,
+      mapName = mapName,
     )
     renderThread = MapboxRenderThread(
       mapboxRenderer = this,
       mapboxWidgetRenderer = widgetRenderer,
       translucentSurface = false,
       antialiasingSampleCount = antialiasingSampleCount,
-      contextMode = contextMode
+      contextMode = contextMode,
+      mapName = mapName,
     )
   }
 
   @VisibleForTesting(otherwise = VisibleForTesting.PRIVATE)
-  internal constructor(renderThread: MapboxRenderThread) {
+  internal constructor(renderThread: MapboxRenderThread) : super(mapName = "") {
     widgetRenderer = MapboxWidgetRenderer(
       antialiasingSampleCount = 1,
+      mapName = "",
     )
     this.renderThread = renderThread
   }